+++ CHANGELOG.md
@@ -50,6 +50,7 @@ This upgrade requires running `evennia migrate` on your existing database
 - [Fix][pull3768]: Make sure the `CmdCopy` command copies object categories,
   since otherwise plurals were lost (jaborsh)
 - [Fix][issue3788]: `GLOBAL_SCRIPTS.all()` raised error (Griatch)
+- [Fix][issue3790]: Fix migration issue due to new db init-check code in launcher (Griatch)
 - Fix: `options` setting `NOPROMPTGOAHEAD` was not possible to set (Griatch)
 - Fix: Make `\\` properly preserve one backlash in funcparser (Griatch)
 - Fix: The testing 'echo' inputfunc didn't work correctly; now returns both args/kwargs (Griatch)
@@ -90,6 +91,7 @@ This upgrade requires running `evennia migrate` on your existing database
 [issue3688]: https://github.com/evennia/evennia/issues/3688
 [issue3687]: https://github.com/evennia/evennia/issues/3687
 [issue3788]: https://github.com/evennia/evennia/issues/3788
+[issue3790]: https://github.com/evennia/evennia/issues/3790